+++ src/json2typescript/any.js
@@ -6,4 +6,4 @@ var Any = (function () {
     return Any;
 }());
 exports.Any = Any;
-//# sourceMappingURL=C:/Utilisateurs/A599669/Documents/GitHub/json2typescript/src/json2typescript/any.js.map
+//# sourceMappingURL=/Users/stephanechevreux/projects/json2typescript/src/json2typescript/any.js.map