+++ src/json2typescript/json-convert-enums.js
@@ -14,4 +14,4 @@ var ValueCheckingMode;
     ValueCheckingMode[ValueCheckingMode["DISALLOW_NULL"] = 3] = "DISALLOW_NULL";
 })(ValueCheckingMode = exports.ValueCheckingMode || (exports.ValueCheckingMode = {}));
 ;
-//# sourceMappingURL=C:/Utilisateurs/A599669/Documents/GitHub/json2typescript/src/json2typescript/json-convert-enums.js.map
+//# sourceMappingURL=/Users/stephanechevreux/projects/json2typescript/src/json2typescript/json-convert-enums.js.map