+++ src/json2typescript/json-convert.ts
@@ -429,7 +429,7 @@ export class JsonConvert {
 
         // Map the property
         try {
-            json[jsonKey] = customConverter !== null ? customConverter.serialize(classInstancePropertyValue) : this.verifyProperty(expectedJsonType, classInstancePropertyValue, true);
+            json[jsonKey] = customConverter !== null ? customConverter.serialize(classInstancePropertyValue, instance) : this.verifyProperty(expectedJsonType, classInstancePropertyValue, true);
         } catch (e) {
             throw new Error(
                 "Fatal error in JsonConvert. " +
@@ -485,7 +485,7 @@ export class JsonConvert {
 
         // Map the property
         try {
-            instance[classPropertyName] = customConverter !== null ? customConverter.deserialize(jsonValue) : this.verifyProperty(expectedJsonType, jsonValue);
+            instance[classPropertyName] = customConverter !== null ? customConverter.deserialize(jsonValue, json) : this.verifyProperty(expectedJsonType, jsonValue);
         } catch (e) {
             throw new Error(
                 "Fatal error in JsonConvert. " +