+++ src/json2typescript/json-custom-convert.d.ts
@@ -1,4 +1,4 @@
 export interface JsonCustomConvert<T> {
-    serialize(data: T): any;
-    deserialize(data: any): T;
+    serialize(data: T, parent?: any): any;
+    deserialize(data: any, parent?: any): T;
 }